Add more submodules.
These submodules should be the essential ones. Stopped using Command-T
and started using Ctrl-P. Command-T is is very good, but it was
conflicting a lot with RVM.

Another one that deserves mention is vim-sensible. This one was not added
as a submodule: it is simple enough to be copy/pasted and I don't want
it to mess with my vimrc when it gets updated.