Finish vimrc.
I don't consider vimrc to be legacy anymore. Seems to be working
properly, with very few testing commands. And even those are simple
enough and shouldn't cause problems.

Something that deserves mention is the leader mapping and modification
of vim-sensible <C-L> mapping. This was necessary since it was
conflicting with other more coherent mappings.