Fix tasks that install asdf plugins
Even though the installation of asdf-vm worked properly, tasks that use
the asdf command failed because $PATH had not been "updated" in the
current terminal.

The fix here is to just source asdf.sh. Note that sourcing .bashrc
instead of asdf.sh may not work because .bashrc may do nothing if not
running interactively.