Parameterize task that edits fstab
The playbook now "detects" on which computer it's running and defines
reliable and convenient variables, which can be used to parameterize
tasks.

The first task that uses this is the recently added one that edits the
fstab: the role includes the variable file specific to my notebook,
which contains the specific device UUID to be mounted.