Fix bug introduced by 35f9a720b438f7acd567e28fe0b9e52901d10522
Although vc-git is a feature (meaning it can be required), it is not a
package, thus it can't be installed with package-install. I've
configured use-package to always ensure packages, therefore an error
occurred because use-package was indeed attempting to install vc-git.

An alternative solution would be to use:

(use-pacakge vc-git
  :ensure vc
  ...)

This way, use-package would supposedly ensure the package vc, and
load the feature with (require 'vc-git).

Anyway, I like to use magit, and my usage of vc-git has been minimal if
not nonexistent, therefore I chose a simpler solution.