Fix bug caused by misuse of tar
By default, tar preserves the original mtime of the files, which implies
that the compressed archive will always have a newer mtime than the
files contained in it. Therefore, targets that have tar files as
prerequisite will always be rebuilt.

The solution is to pass the "--touch" flag to tar.